Initial commit
Initial commit of Balrog, an implementation of the `pass` password
manager using only a POSIX shell and utilities. This is designed
to support a subset of the features of `pass` while remaining
100% compatible with its API so that it can be used as a drop-in
replacement.

Implemented so far are:
- commands:
	- edit
	- generate
	- show
- flags:
	- -c/--clip
	- -i/--in-place
	- -n/--no-symbols

Also included is a *very* skinny "otp" command that covers some
of the functionality of pass-otp. This is by no means complete
and doesn't support much outside of my specific use case. Not sure
if it's eventually worth reimplementing pass-otp myself as an
extension or what.