Remove the temporary file regardless of whether the gpg command fails
Previously the temp file could have been left in place if the gpg
command failed for some reason. Now, if it exists, it should always
get cleaned up.

There remains one scenario where it might be left dangling, and
that's if the whole balrog process gets killed. I'm not sure what
to do about that scenario. I could spawn a cleanup process that
would delete the file after balrog exits, but you could still just
hard power off the computer and be left with a dangling file. My
current thinking is just to leave it for the time being --- it's
in the user's home folder, so it's not readable by any other users,
and it'll get overwritten next time balrog is invoked.